refactor: :art: Reorganize components
