+++ ui/src/pages/invoices/components/InvoiceForm.tsx
@@ -1,9 +1,9 @@
-import { Invoice } from "./types";
-import { useFetchClients } from "../clients/hooks";
-import { useSubmitInvoice } from "./hooks";
-import SubmissionForm from "../../common/components/SubmissionForm";
-import { mapGetClientToClient } from "../clients/mapper";
-import { mapInvoicetoCreateInvoice } from "./mappers";
+import { Invoice } from "../types";
+import { useFetchClients } from "../../clients/hooks";
+import { useSubmitInvoice } from "../hooks";
+import SubmissionForm from "../../../common/components/SubmissionForm";
+import { mapGetClientToClient } from "../../clients/mapper";
+import { mapInvoicetoCreateInvoice } from "../mappers";
 import { Alert } from "react-bootstrap";
 
 type InvoiceFormData = Omit<Invoice, "id">;
